feat: image uploads
This adds support for the undocumented image upload API endpoint that's
not part of the official API.

The `APIRequest` type adds a `URL` field that can be set to explicitly
set a URL, as in the case with images since the endpoint is completely
different.

This only supports uploading images.